patchpanel: build with -Wconversion
This patch fixes various errors and implicit casts in patchpanel
packages detected with the -Wconversion compiler flag.

This covers the following ebuild packages: patchpanel,
patchpanel-client, mcastd, ndproxyd.

BUG=b:236092161
TEST=FEATURES=test emerge-$BOARD patchpanel patchpanel-client ndproxyd \
  mcastd, with $BOARD==atlas (x86) and $BOARD==trogdor (arm).
Also built patchpanel and patchpanel-client fuzzers on amd64-generic
with asan fuzzer profile.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3857509
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>

+++ patchpanel/address_manager.cc
@@ -67,7 +67,7 @@ AddressManager::AddressManager() {
   }
 }
 
-MacAddress AddressManager::GenerateMacAddress(uint8_t index) {
+MacAddress AddressManager::GenerateMacAddress(uint32_t index) {
   return index == kAnySubnetIndex ? mac_addrs_.Generate()
                                   : mac_addrs_.GetStable(index);
 }
