+++ patchpanel/dns/big_endian.h
@@ -20,7 +20,7 @@ namespace base {
 // This would cause SIGBUS on ARMv5 or earlier and ARMv6-M.
 template <typename T>
 inline void ReadBigEndian(const char buf[], T* out) {
-  *out = buf[0];
+  *out = static_cast<uint8_t>(buf[0]);
   for (size_t i = 1; i < sizeof(T); ++i) {
     *out <<= 8;
     // Must cast to uint8_t to avoid clobbering by sign extension.
@@ -41,7 +41,7 @@ inline void WriteBigEndian(char buf[], T val) {
 // Specializations to make clang happy about the (dead code) shifts above.
 template <>
 inline void ReadBigEndian<uint8_t>(const char buf[], uint8_t* out) {
-  *out = buf[0];
+  *out = static_cast<uint8_t>(buf[0]);
 }
 
 template <>
@@ -56,7 +56,7 @@ class BASE_EXPORT BigEndianReader {
   BigEndianReader(const char* buf, size_t len);
 
   const char* ptr() const { return ptr_; }
-  int remaining() const { return end_ - ptr_; }
+  ssize_t remaining() const { return end_ - ptr_; }
 
   bool Skip(size_t len);
   bool ReadBytes(void* out, size_t len);
@@ -83,7 +83,7 @@ class BASE_EXPORT BigEndianWriter {
   BigEndianWriter(char* buf, size_t len);
 
   char* ptr() const { return ptr_; }
-  int remaining() const { return end_ - ptr_; }
+  ssize_t remaining() const { return end_ - ptr_; }
 
   bool Skip(size_t len);
   bool WriteBytes(const void* buf, size_t len);
